flag

+++ pol-core/pol/ufunc.cpp
@@ -1996,14 +1996,13 @@ void send_feature_enable( Client* client )
   {
     if ( Plib::systemstate.config.character_slots == 7 )
     {
-      clientflag |=
-          static_cast<u32>( B9Feature::Has7thSlot );  // 7th & 6th character flag (B9 Packet)
-      clientflag &= ~static_cast<u32>( B9Feature::ThirdDawn );  // Disable Third Dawn?
+      clientflag |= B9Feature::Has7thSlot;  // 7th & 6th character flag (B9 Packet)
+      clientflag &= ~B9Feature::ThirdDawn;  // Disable Third Dawn?
     }
     else if ( Plib::systemstate.config.character_slots == 6 )
     {
-      clientflag |= static_cast<u32>( B9Feature::Has6thSlot );  // 6th character flag (B9 Packet)
-      clientflag &= ~static_cast<u32>( B9Feature::ThirdDawn );
+      clientflag |= B9Feature::Has6thSlot;  // 6th character flag (B9 Packet)
+      clientflag &= ~B9Feature::ThirdDawn;
     }
   }
 
@@ -2011,7 +2010,7 @@ void send_feature_enable( Client* client )
   if ( client->UOExpansionFlag & KR )
   {
     if ( settingsManager.ssopt.support_faces == 2 )
-      clientflag |= static_cast<u32>( B9Feature::KRFaces );
+      clientflag |= B9Feature::KRFaces;
   }
 
   PktHelper::PacketOut<PktOut_B9> msg;
